Remove old Add role wizard
