+++ src/smart-components/role/add-role-permissions/add-role-permission-wizard.js
@@ -10,8 +10,8 @@ import { updateRole } from '../../../redux/actions/role-actions.js';
 import FormRenderer from '@data-driven-forms/react-form-renderer/form-renderer';
 import Pf4FormTemplate from '@data-driven-forms/pf4-component-mapper/form-template';
 import componentMapper from '@data-driven-forms/pf4-component-mapper/component-mapper';
-import AddPermissionsTable from '../add-role-new/add-permissions';
-import CostResources from '../add-role-new/cost-resources';
+import AddPermissionsTable from '../add-role/add-permissions';
+import CostResources from '../add-role/cost-resources';
 import { schemaBuilder } from './schema';
 import { useIntl } from 'react-intl';
 import messages from '../../../Messages';
