+++ src/smart-components/role/add-role/add-permissions-template.js
