+++ src/smart-components/role/add-role/add-role-success.js
