+++ src/smart-components/role/add-role/add-role-wizard.js
@@ -1,163 +1,218 @@
-import React, { useState } from 'react';
-import { shallowEqual, useDispatch, useSelector } from 'react-redux';
-import { useHistory } from 'react-router-dom';
+import React, { useState, useEffect, useRef, createContext } from 'react';
 import PropTypes from 'prop-types';
-import { Wizard } from '@patternfly/react-core';
+import { useDispatch } from 'react-redux';
 import { addNotification } from '@redhat-cloud-services/frontend-components-notifications/';
-import { createRole, fetchRoles } from '../../../redux/actions/role-actions';
-import SummaryContent from './summary-content';
-import ResourceDefinitions from './resource-definitions';
-import RoleInformation from './role-information';
-import PermissionInformation from './permission-information';
+import FormRenderer from '@data-driven-forms/react-form-renderer/form-renderer';
+import Pf4FormTemplate from '@data-driven-forms/pf4-component-mapper/form-template';
+import componentMapper from '@data-driven-forms/pf4-component-mapper/component-mapper';
+import { Wizard } from '@patternfly/react-core';
+import { schemaBuilder } from './schema';
+import { createRole, fetchRolesWithPolicies } from '../../../redux/actions/role-actions';
 import { WarningModal } from '../../common/warningModal';
+import AddRoleSuccess from './add-role-success';
+import BaseRoleTable from './base-role-table';
+import AddPermissionsTable from './add-permissions';
+import ReviewStep from './review';
+import CostResources from './cost-resources';
+import TypeSelector from './type-selector';
+import { useHistory } from 'react-router-dom';
+import { createQueryParams } from '../../../helpers/shared/helpers';
+import paths from '../../../utilities/pathnames';
 import { useIntl } from 'react-intl';
 import messages from '../../../Messages';
-import '../../common/hideWizard.scss';
 
-const AddRoleWizard = () => {
-  const intl = useIntl();
-  const [formData, setFormData] = useState({});
-  const [isRoleFormValid, setIsRoleFormValid] = useState(false);
-  const [isPermissionFormValid, setIsPermissionFormValid] = useState(false);
-  const [stepIdReached, setStepIdReached] = useState(1);
-  const { push } = useHistory();
+import './add-role-wizard.scss';
+import SilentErrorBoundary from '../../common/silent-error-boundary';
 
-  const { pagination } = useSelector(
-    ({ roleReducer: { roles, filterValue, isLoading } }) => ({
-      roles: roles.data,
-      pagination: roles.meta,
-      isLoading,
-      searchFilter: filterValue,
-    }),
-    shallowEqual
-  );
+export const AddRoleWizardContext = createContext({
+  success: false,
+  submitting: false,
+  error: undefined,
+});
 
-  const dispatch = useDispatch();
+const FormTemplate = (props) => <Pf4FormTemplate {...props} showFormControls={false} />;
 
-  const handleChange = (data) => {
-    setFormData({ ...formData, ...data });
-  };
-
-  const handleRoleChange = (data, isValid) => {
-    handleChange(data);
-    setIsRoleFormValid(isValid);
-  };
-
-  const handlePermissionChange = (data, isValid) => {
-    handleChange(data);
-    setIsPermissionFormValid(isValid);
-  };
-
-  const steps = [
-    {
-      id: 1,
-      name: intl.formatMessage(messages.nameAndDescription),
-      canJumpTo: stepIdReached >= 1,
-      component: new RoleInformation(formData, handleRoleChange),
-      enableNext: isRoleFormValid,
-    },
-    {
-      id: 2,
-      name: intl.formatMessage(messages.permission),
-      canJumpTo: stepIdReached >= 2 && isRoleFormValid,
-      component: new PermissionInformation(formData, handlePermissionChange),
-      enableNext: isPermissionFormValid,
-    },
-    {
-      id: 3,
-      name: intl.formatMessage(messages.resourceDefinitions),
-      canJumpTo: stepIdReached >= 3 && isRoleFormValid && isPermissionFormValid,
-      component: new ResourceDefinitions(formData, handleChange),
-    },
-    {
-      id: 4,
-      name: intl.formatMessage(messages.review),
-      canJumpTo: stepIdReached >= 4 && isRoleFormValid && isPermissionFormValid,
-      component: new SummaryContent(formData),
-      nextButtonText: intl.formatMessage(messages.confirm),
-    },
-  ];
-
-  const onNext = ({ id }) => {
-    const step = stepIdReached < id ? id : stepIdReached;
-    setStepIdReached(step);
-  };
-
-  const onSubmit = async () => {
-    const roleData = {
-      applications: [formData.application],
-      description: formData.description,
-      name: formData.name,
-      access: [
-        {
-          // Permission must be in the format "application:resource_type:operation"
-          permission: `${formData.application}:${formData.resourceType}:${formData.permission}`,
-          resourceDefinitions: formData.resourceDefinitions.map((definition) => {
-            return {
-              attributeFilter: {
-                key: definition.key,
-                operation: definition.operation,
-                value: definition.value,
-              },
-            };
-          }),
-        },
-      ],
-    };
-    const role = await dispatch(createRole(roleData));
-    dispatch(fetchRoles(pagination).then(push('/roles')));
-    return role;
-  };
-
-  const onCancel = () => {
-    dispatch(
-      addNotification({
-        variant: 'warning',
-        title: intl.formatMessage(messages.creatingRoleCanceled),
-        dismissDelay: 8000,
-      })
-    );
-    push('/roles');
-  };
-
-  const [cancelWarningVisible, setCancelWarningVisible] = useState(false);
-
-  return (
-    <React.Fragment>
-      <Wizard
-        className={cancelWarningVisible && 'rbac-m-wizard__hidden'}
-        title={intl.formatMessage(messages.addRole)}
-        isOpen
-        onClose={() => {
-          (!Object.values(formData).filter(Boolean).length > 0 && onCancel()) || setCancelWarningVisible(true);
-        }}
-        onNext={onNext}
-        onSave={onSubmit}
-        steps={steps}
-      />
-      <WarningModal type="role" isOpen={cancelWarningVisible} onModalCancel={() => setCancelWarningVisible(false)} onConfirmCancel={onCancel} />
-    </React.Fragment>
-  );
+const Description = ({ Content, ...rest }) => <Content {...rest} />;
+Description.propTypes = {
+  Content: PropTypes.elementType.isRequired,
 };
 
-AddRoleWizard.defaultProps = {
-  users: [],
-  inputValue: '',
-  selectedGroup: undefined,
-  selectedUsers: [],
-  selectedRoles: [],
+export const mapperExtension = {
+  'base-role-table': BaseRoleTable,
+  'add-permissions-table': AddPermissionsTable,
+  'cost-resources': CostResources,
+  review: ReviewStep,
+  description: Description,
+  'type-selector': TypeSelector,
+};
+
+const AddRoleWizard = ({ pagination, filters, orderBy }) => {
+  const intl = useIntl();
+  const dispatch = useDispatch();
+  const { push } = useHistory();
+  const [wizardContextValue, setWizardContextValue] = useState({
+    success: false,
+    submitting: false,
+    error: undefined,
+    hideForm: false,
+  });
+  const [cancelWarningVisible, setCancelWarningVisible] = useState(false);
+  const container = useRef(document.createElement('div'));
+  const [schema, setSchema] = useState();
+
+  useEffect(() => {
+    setSchema(schemaBuilder(container.current));
+  }, []);
+
+  const onClose = () =>
+    push({
+      pathname: paths.roles.path,
+      search: createQueryParams({ page: 1, per_page: pagination.limit }),
+    });
+
+  const onCancel = () => {
+    if (!wizardContextValue.success) {
+      dispatch(
+        addNotification({
+          variant: 'warning',
+          title: intl.formatMessage(messages.creatingRoleCanceled),
+          dismissDelay: 8000,
+        })
+      );
+    }
+
+    setCancelWarningVisible(false);
+    /**
+     * This timeout should force React to wait for the modal close and push to history afterwards.
+     * That should fix the runtime error we are seeing in the production version of the code.
+     */
+    setTimeout(() => {
+      push({
+        pathname: paths.roles.path,
+        search: createQueryParams({ page: 1, per_page: pagination.limit, ...filters }),
+      });
+    });
+  };
+
+  const setWizardError = (error) => setWizardContextValue((prev) => ({ ...prev, error }));
+  const setWizardSuccess = (success) => setWizardContextValue((prev) => ({ ...prev, success }));
+  const setHideForm = (hideForm) => setWizardContextValue((prev) => ({ ...prev, hideForm }));
+
+  const onSubmit = (formData) => {
+    const {
+      'role-name': name,
+      'role-description': description,
+      'role-copy-name': copyName,
+      'role-copy-description': copyDescription,
+      'add-permissions-table': permissions,
+      'cost-resources': resourceDefinitions,
+      'role-type': type,
+    } = formData;
+    setWizardContextValue((prev) => ({ ...prev, submitting: true }));
+
+    const selectedPermissionIds = permissions.map((record) => record.uuid);
+    const roleData = {
+      applications: [...new Set(permissions.map(({ uuid: permission }) => permission.split(':')[0]))],
+      description: (type === 'create' ? description : copyDescription) || null,
+      name: type === 'create' ? name : copyName,
+      access: permissions.reduce(
+        (acc, { uuid: permission, requires = [] }) => [
+          ...acc,
+          ...[permission, ...requires.filter((require) => !selectedPermissionIds.includes(require))].map((permission) => ({
+            permission,
+            resourceDefinitions: resourceDefinitions?.find((r) => r.permission === permission)
+              ? [
+                  {
+                    attributeFilter: {
+                      key: `cost-management.${permission.split(':')[1]}`,
+                      operation: 'in',
+                      value: resourceDefinitions?.find((r) => r.permission === permission).resources,
+                    },
+                  },
+                ]
+              : [],
+          })),
+        ],
+        []
+      ),
+    };
+    return dispatch(createRole(roleData))
+      .then(() => {
+        setWizardContextValue((prev) => ({ ...prev, submitting: false, success: true, hideForm: true }));
+        dispatch(fetchRolesWithPolicies({ limit: pagination.limit, orderBy, inModal: false }));
+      })
+      .catch(() => {
+        setWizardContextValue((prev) => ({ ...prev, submitting: false, success: false, hideForm: true }));
+        dispatch(fetchRolesWithPolicies({ limit: pagination.limit, orderBy, inModal: false }));
+        onClose();
+      });
+  };
+
+  if (!schema) {
+    return null;
+  }
+  return (
+    <AddRoleWizardContext.Provider value={{ ...wizardContextValue, setWizardError, setWizardSuccess, setHideForm }}>
+      <SilentErrorBoundary silentErrorString="focus-trap">
+        <WarningModal
+          type="role"
+          isOpen={cancelWarningVisible}
+          onModalCancel={() => {
+            container.current.hidden = false;
+            setCancelWarningVisible(false);
+          }}
+          onConfirmCancel={onCancel}
+        />
+      </SilentErrorBoundary>
+      {wizardContextValue.hideForm ? (
+        wizardContextValue.success ? (
+          <Wizard
+            title={intl.formatMessage(messages.createRole)}
+            isOpen
+            onClose={onClose}
+            steps={[
+              {
+                name: 'success',
+                component: <AddRoleSuccess onClose={onClose} />,
+                isFinishedStep: true,
+              },
+            ]}
+          />
+        ) : null
+      ) : (
+        <FormRenderer
+          schema={schema}
+          container={container}
+          subscription={{ values: true }}
+          FormTemplate={FormTemplate}
+          initialValues={{
+            'role-type': 'create',
+          }}
+          componentMapper={{ ...componentMapper, ...mapperExtension }}
+          onSubmit={onSubmit}
+          onCancel={(values) => {
+            const showWarning = Boolean((values && values['role-name']) || values['role-description'] || values['copy-base-role']);
+            if (showWarning) {
+              container.current.hidden = true;
+              setCancelWarningVisible(true);
+            } else {
+              onCancel();
+            }
+          }}
+        />
+      )}
+    </AddRoleWizardContext.Provider>
+  );
 };
 
 AddRoleWizard.propTypes = {
-  addNotification: PropTypes.func.isRequired,
-  createRole: PropTypes.func.isRequired,
-  fetchRoles: PropTypes.func.isRequired,
-  inputValue: PropTypes.string,
   pagination: PropTypes.shape({
     limit: PropTypes.number.isRequired,
-    offset: PropTypes.number.isRequired,
-    count: PropTypes.number.isRequired,
-  }),
+  }).isRequired,
+  filters: PropTypes.shape({
+    name: PropTypes.string,
+  }).isRequired,
+  orderBy: PropTypes.string,
 };
 
 export default AddRoleWizard;
