+++ src/smart-components/role/add-role/cost-resources-template.js
