+++ src/smart-components/role/add-role/cost-resources.js
